aider: Added Ansible role to install MongoDB on openSUSE Tumbleweed.
# Aider chat conversation:

USER: Could you please create an ansible role to install mongodb on an opensuse tumbleweed install

ASSISTANT: Sure, I can help with that. We will need to create a new Ansible role for installing MongoDB on openSUSE Tumbleweed.

The role will be defined in the `roles/mongodb/tasks/opensuse.yml` file.

Here is the Ansible role:

```yaml
roles/mongodb/tasks/opensuse.yml
<<<<<<< HEAD

=======
- name: Import the MongoDB public GPG Key
  rpm_key:
    key: https://www.mongodb.org/static/pgp/server-4.4.asc
    state: present

- name: Add the MongoDB Repository
  get_url:
    url: https://repo.mongodb.org/yum/redhat/7/mongodb-org/4.4/x86_64/
    dest: /etc/yum.repos.d/mongodb-org-4.4.repo
    mode: '0644'

- name: Install MongoDB packages
  zypper:
    name: mongodb-org
    state: present

- name: Enable MongoDB service
  systemd:
    name: mongod
    enabled: yes

- name: Start MongoDB service
  systemd:
    name: mongod
    state: started
>>>>>>> updated
```

This Ansible role does the following:

1. Imports the MongoDB public GPG Key.
2. Adds the MongoDB Repository.
3. Installs the MongoDB packages.
4. Enables the MongoDB service to start on boot.
5. Starts the MongoDB service.

Please note that this is a basic role and does not include any hardening or optimization configurations. You might want to adjust it according to your needs.