aider: Ensure python3, python3-pip and universal-ctags are installed.
# Aider chat conversation:

USER: Please update this role to also install universal-ctags on an opensuse tumbleweed install

ASSISTANT: To install universal-ctags on an opensuse tumbleweed install, we need to add a new task in the `roles/aider/tasks/opensuse.yml` file. This task will use the `zypper` module to ensure that the `universal-ctags` package is installed.

Here is the *edit block* for the change:

```yaml
roles/aider/tasks/opensuse.yml
<<<<<<< HEAD
- name: Ensure python3 and python3-pip are installed
  zypper:
    name:
      - python3
      - python3-pip
    state: present

- name: Install aider-chat Python package
  pip:
    name: aider-chat
    executable: pip3
=======
- name: Ensure python3, python3-pip and universal-ctags are installed
  zypper:
    name:
      - python3
      - python3-pip
      - universal-ctags
    state: present

- name: Install aider-chat Python package
  pip:
    name: aider-chat
    executable: pip3
>>>>>>> updated
```

This change will ensure that the `universal-ctags` package is installed along with `python3` and `python3-pip`.